use dedicated header for QgsAttributes
