+++ python/core/core.sip
@@ -21,6 +21,7 @@
 %Include qgsactionscoperegistry.sip
 %Include qgsaggregatecalculator.sip
 %Include qgsanimatedicon.sip
+%Include qgsattributes.sip
 %Include qgsattributetableconfig.sip
 %Include qgsattributeeditorelement.sip
 %Include qgsbearingutils.sip
